backend/requirements.txt
fastapi
uvicorn[standard]
dotenv
pydantic
python-dotenv
google-genai
jsonify
smtplib
email
pydantic[email]